Removed showcase app.
After a lot of thinking about the project, I think it will be best to
focus on getting our users onto featured sites as easily as possible. I
feel that currently, the showcase app seems at odds with that user's
mission.

This commit reflects the return to that core user story.