Fixed bug with window call.
Initially I had figured that window would be undefined and I could just
use a check for it. Turns out you need to do a try/except style catch
for it, but furthermore React ends up in a funky state.

I found that React's virtual DOM ended up out of sync with the actual
DOM. The solution I arrived on is using state instead of props, and
updating state in a componentDidMount() lifecycle method instead of in
the render method.